pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ateam"
version = "0.1.0"
description = "A-TEAM: Agentic development system with orchestrated AI agents"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "tomli>=2.0; python_version < '3.11'",
    "rich>=13.0",
    "pydantic>=2.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "sse-starlette>=1.6",
]

[project.scripts]
ateam = "ateam.cli:main"

[tool.setuptools.packages.find]
include = ["ateam*"]
